requirements.txt
click==8.1.7
dateparser==1.2.0
greenlet==3.0.3
notify2==0.3.1
prettytable==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzlocal==5.2
wcwidth==0.2.13
